Add codebase for kairos project
Includes a java-based interpreter for a Whitespace assembler
dialect created for easier reading and writing.
The interpreter is the only currently available interpreter to be
fully compliant with the 0.3 specification of whitespace
Also allows translation from genuine Whitespace code to the
WhitespaceAssembler dialect and back again.
Interactive session and weaving of code into existing text files
with non-whitespace content are stubs only.